Finished updating UI to better reflect UI diagram.
- The mock UI has tabs to make implementation a little cleaner. The final
UI design might be updated to include tabs, but it currently uses
buttons to switch between the task detail screen and the timer screen.
In the future, there will be a task list screen instead, and clicking a
task in the task list will bring up the detail screen.

- TaskStore now accepts a default task name string instead of introducing
dependencies on ActivityListener.

- TaskStore and Task now have a listener/subscriber interface to allow the
presenter instances to monitor changes to certain tasks, or complete
swaps of the current task.

- Unit tests for all the new Task/TaskStore features, tests for the timer and
setuptask presenters.